chore: move decorator out of src
